% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. If \code{TRUE}, then missing values are removed.}
}
\value{
The standard error of the mean.
}
\description{
This function computes the standard error of the values in \code{x}. If \code{na.rm} is \code{TRUE} then missing values are removed before computation proceeds.
}
\examples{
## Usage with and without NAs
se(c(1, 2, 5, 7))
se(c(1, 2, 5, NA, 7))
se(c(1, 2, 5, NA, 7), na.rm = TRUE)
}
